+++ client/src/components/EventPreview.jsx
