+++ client/src/components/GTSlider.jsx
