+++ client/src/components/ParticipantItems.jsx
