+++ client/src/components/bracket/DisplayBracket.jsx
@@ -1,0 +1,7 @@
+export default function DisplayBracket(BracketInfo, Participants = []){
+    return(
+      <div id='bracket-display'> 
+
+      </div>
+    );
+}
