+++ client/src/pages/FAQ.jsx
